[benchmarks/pypsa] Simplify config files and shrink benchmarks (#35)
Hi @siddharth-krishna , in the last commit I removed those lines that
`pypsa-eur-elec-10-lvopt-3h.yaml` has in common with the
`common.default.yaml` file for pypsa v 0.11.0 (which is the one this
config file refers too). However, apart from reducing the size of the
file I'm not so convinced of the reason why we should need to cut parts
of this file. I guess we'll be having a reference `config.default.yaml`
file and an additional script to modify it according to the specific
info related to each of the pypsa-eur-based sample problems, but I can't
see why we should be doing that, if so.

---------

Co-authored-by: Siddharth Krishna <siddharth-krishna@users.noreply.github.com>
Co-authored-by: jacek-oet <jacek.bendig@openenergytransition.org>