Re-run GLPK results after fixing linopy MPS format (#86)
The errors we were seeing with GLPK in #68 was because linopy by default
calls `glpsol` with the `--mps` argument, which is for the fixed MPS
format, but the JuMP-HiGHS benchmarks are using the free MPS format
(`--freemps`), so I re-ran the GLPK runs by hacking linopy to use
`--freemps`:
```diff
diff --git a/linopy/solvers.py b/linopy/solvers.py
index e89a6a9..2c996e3 100644
--- a/linopy/solvers.py
+++ b/linopy/solvers.py
@@ -563,7 +563,8 @@ class GLPK(Solver):
         Path(solution_fn).parent.mkdir(exist_ok=True)
 
         # TODO use --nopresol argument for non-optimal solution output
-        command = f"glpsol --{io_api} {problem_fn} --output {solution_fn} "
+        io_api_arg = "freemps" if io_api == "mps" else io_api
+        command = f"glpsol --{io_api_arg} {problem_fn} --output {solution_fn} "
         if log_fn is not None:
             command += f"--log {log_fn} "
         if warmstart_fn:
```

This PR has the new results, and I'll open a linopy PR to fix the issue
in a more robust way in the new year. (FYI @FabianHofmann )

There are still a few benchmarks returning `uknown` statuses, this needs
to be looked into.