Benchmarks: fix URLs to use name-size filename convention (#337)
The website and the benchmark runner uses the format
`{name}-{size}.{lp/mps}` for benchmark files, but some files on GCS
buckets use different file names.

This PR renames everything on the bucket to follow the
`{name}-{size}.{lp/mps}` format as well, and also updates the URLs in
the `metadata.yaml` files.

I have also renamed the files on GCS accordingly.