Benchmarks: add ethos.fine multi regional example (#333)
This is the new PR, as mentioned in #329. I added the requested
description of the regional focus of the example.
I also reviewed the other examples. However, most of them are very
simple scripts designed to demonstrate how the fine code works. I don't
think they're complex enough to serve as a benchmark.

---------

Co-authored-by: Daniele Lerede <daniele.lerede@openenergytransition.org>
Co-authored-by: Siddharth Krishna <siddharth-krishna@users.noreply.github.com>
Co-authored-by: Jacek <jacek.bendig@openenergytransition.org>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>